pyproject.toml
[tool.poetry]
name = "socialwebapp"
version = "0.1.0"
description = ""
authors = ["RocketMenace <grizzlygit777@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
ruff = "^0.9.4"
isort = "^6.0.0"
pydantic-settings = "^2.7.1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.37"}
asyncpg = "^0.30.0"
fastapi = "^0.115.8"
uvicorn = "^0.34.0"
uvloop = "^0.21.0"
httptools = "^0.6.4"
psycopg2-binary = "^2.9.10"
passlib = "^1.7.4"
pytest = "^8.3.4"
httpx = "^0.28.1"
trio = "^0.28.0"
python-multipart = "^0.0.20"
aiofiles = "^24.1.0"
pyjwt = "^2.10.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
